// frontend/src/app/core/models/cuenta.model.ts
import { Bolsillo } from './bolsillo.model';

export interface Cuenta {
  _id?: string;
  id?: string;
  numero_cuenta: string;
  color: string;
  saldo: number;
  meta_ahorro: number;
  tipo: string;
  createdAt?: Date;
  usuario_id?: string;
  id_bolsillo?: string | null;

  usuarioNombre?: string;
  tieneBolsillo?: boolean;
  bolsillos?: Bolsillo[];
}